commit view

+++ database/migrations/2024_07_19_103146_create_order_items_table.php
@@ -18,9 +18,9 @@ return new class extends Migration
             $table->foreignIdFor(Order::class)->constrained();
             $table->foreignIdFor(Product::class)->constrained();
             $table->string('name');
-            $table->float('price');
+            $table->decimal('price', 10, 2);
             $table->integer('quantity');
-            $table->float('total');
+            $table->decimal('total', 15, 2);
             $table->timestamps();
         });
     }
